proto: move DigestType into dnssec root module
